+++ batch/src/main/java/com/ImSnacks/NyeoreumnagiBatch/common/enums/WeatherCondition.java
@@ -11,7 +11,7 @@ public enum WeatherCondition {
     SNOW("눈"),
     NIGHT("맑음"),
     CLOUDY_NIGHT("흐림"),
-    HEAT_WAVE("폭염"),
+    MIDDLE("보통")
     ;
 
     private final String keyword;
@@ -22,5 +22,35 @@ public enum WeatherCondition {
 
     public String getKeyword() {
         return keyword;
+    }
+
+    public static WeatherCondition fromApiString(String apiValue) {
+        if (apiValue == null) return null;
+
+        if (apiValue.equals("맑음")) {
+            return SUNNY;
+        }
+
+        if (apiValue.contains("구름많")) {
+            if (apiValue.contains("비") || apiValue.contains("소나기")) {
+                return RAIN;
+            }
+            if (apiValue.contains("눈")) {
+                return SNOW;
+            }
+            return LESS_CLOUDY;
+        }
+
+        // 흐림 계열
+        if (apiValue.contains("흐림") || apiValue.contains("흐리")) {
+            if (apiValue.contains("비") || apiValue.contains("소나기")) {
+                return RAIN;
+            }
+            if (apiValue.contains("눈")) {
+                return SNOW;
+            }
+            return CLOUDY;
+        }
+        return MIDDLE;
     }
 }
