+++ batch/src/main/java/com/ImSnacks/NyeoreumnagiBatch/seven_days_weather_forecast/dto/SevenDayWeatherConditionDto.java
@@ -1,6 +1,6 @@
 package com.ImSnacks.NyeoreumnagiBatch.seven_days_weather_forecast.dto;
 
-import com.ImSnacks.NyeoreumnagiBatch.writer.entity.WeatherCondition;
+import com.ImSnacks.NyeoreumnagiBatch.common.enums.WeatherCondition;
 
 import java.time.LocalDate;
 import java.util.List;
