+++ batch/src/main/java/com/ImSnacks/NyeoreumnagiBatch/seven_days_weather_forecast/writer/SevenDayTemperatureWriter.java
@@ -1,7 +1,7 @@
 package com.ImSnacks.NyeoreumnagiBatch.seven_days_weather_forecast.writer;
 
+import com.ImSnacks.NyeoreumnagiBatch.common.entity.SevenDayWeatherForecast;
 import com.ImSnacks.NyeoreumnagiBatch.seven_days_weather_forecast.dto.SevenDayTemperatureForecastDto;
-import com.ImSnacks.NyeoreumnagiBatch.writer.entity.SevenDayWeatherForecast;
 import com.ImSnacks.NyeoreumnagiBatch.writer.repository.SevenDayWeatherForecastRepository;
 import jakarta.transaction.Transactional;
 import lombok.extern.slf4j.Slf4j;
