+++ batch/src/main/java/com/ImSnacks/NyeoreumnagiBatch/shortTermWeatherForecast/processor/WeatherProcessor.java
@@ -5,6 +5,7 @@ import com.ImSnacks.NyeoreumnagiBatch.shortTermWeatherForecast.processor.dto.Vil
 import com.ImSnacks.NyeoreumnagiBatch.shortTermWeatherForecast.reader.dto.VilageFcstResponseDto;
 import com.ImSnacks.NyeoreumnagiBatch.shortTermWeatherForecast.processor.utils.weatherRiskFilter.WeatherRiskFilter;
 import com.ImSnacks.NyeoreumnagiBatch.shortTermWeatherForecast.writer.dto.ShortTermWeatherDto;
+import lombok.NoArgsConstructor;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.batch.item.ItemProcessor;
